% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAbsRatio.R
\name{calcAbsRatio}
\alias{calcAbsRatio}
\title{calcAbsRatio}
\usage{
calcAbsRatio(
  absorbanceData,
  concentrationData = FALSE,
  wavelength1,
  wavelength2,
  correctFe = FALSE
)
}
\arguments{
\item{absorbanceData}{User input of the table of NEON absorbance data \link{data.frame}}

\item{concentrationData}{User input of the table of NEON concentration data \link{data.frame}}

\item{wavelength1}{User input of the wavelength to be used in the numerator \link{numeric}}

\item{wavelength2}{User input of the wavelength to be used in the denominator \link{numeric}}

\item{correctFe}{User input of whether correction for overlapping absorbption of Fe(III)
should also be included. See README for more information. Defaults to FALSE. \link{logical}}
}
\value{
This function returns a table of absorbance ratio values
}
\description{
This function calculates the absorbance ratio for two user-
specified wavelengths from National Ecological Observatory Network (NEON)
water chemistry data.
}
\examples{
#Using an example file
#outputData <- calcAbsRatio(
#absorbanceData=swc_externalLabAbsorbanceScan,
#concentrationData=swc_externalLabDataByAnalyte,
#wavelength1 = 300
#wavelength2 = 400
#correctFe=TRUE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Robert Hensley \email{hensley@battelleecology.org} \cr
}
\keyword{Absorbance}
\keyword{DOC,}
\keyword{NEON,}
\keyword{SUVA,}
\keyword{UV}
