% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcE2E3.R
\name{calcE2E3}
\alias{calcE2E3}
\title{calcE2E3}
\usage{
calcE2E3(absorbanceData, concentrationData, correctFe = FALSE)
}
\arguments{
\item{absorbanceData}{User input of the table of NEON absorbance data \link{data.frame}}

\item{concentrationData}{User input of the table of NEON concentration data \link{data.frame}}

\item{correctFe}{User input of whether correction for overlapping absorbption of Fe(III)
should also be included. See README for more information. Defaults to FALSE. \link{logical}}
}
\value{
This function returns a table of E2:E3 values
}
\description{
This function calculates the E2:E3 (250 nm : 365 nm) absorbance
ratio from National Ecological Observatory Network (NEON) water
chemistry data.
}
\examples{
#Using an example file
#outputData <- calcE2E3(
#absorbanceData=swc_externalLabAbsorbanceScan,
#concentrationData=swc_externalLabDataByAnalyte,
#correctFe=FALSE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Robert Hensley \email{hensley@battelleecology.org} \cr
}
\keyword{Absorbance}
\keyword{DOC,}
\keyword{NEON,}
\keyword{SUVA,}
\keyword{UV}
