% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSuva350.R
\name{calcSuva350}
\alias{calcSuva350}
\title{calcSuva350}
\usage{
calcSuva350(absorbanceData, concentrationData, correctFe = FALSE)
}
\arguments{
\item{absorbanceData}{User input of the table of NEON absorbance data \link{data.frame}}

\item{concentrationData}{User input of the table of NEON concentration data \link{data.frame}}

\item{correctFe}{User input of whether correction for overlapping absorbption of Fe(III)
should also be included. See README for more information. Defaults to FALSE. \link{logical}}
}
\value{
This function returns a table of SUVA350 values
}
\description{
This function calculates the specific ultra-violet absorbance
(SUVA) at 350 nm from National Ecological Observatory Network (NEON) water
chemistry data.
}
\examples{
#Using an example file
#outputData <- calcSuva350(
#absorbanceData=swc_externalLabAbsorbanceScan,
#concentrationData=swc_externalLabDataByAnalyte,
#correctFe=FALSE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Robert Hensley \email{hensley@battelleecology.org} \cr
}
\keyword{Absorbance}
\keyword{DOC,}
\keyword{NEON,}
\keyword{SUVA,}
\keyword{UV}
