% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSuvaX.R
\name{calcSuvaX}
\alias{calcSuvaX}
\title{calcSuvaX}
\usage{
calcSuvaX(absorbanceData, concentrationData, wavelength, correctFe = FALSE)
}
\arguments{
\item{absorbanceData}{User input of the table of NEON absorbance data \link{data.frame}}

\item{concentrationData}{User input of the table of NEON concentration data \link{data.frame}}

\item{correctFe}{User input of whether correction for overlapping absorbption of Fe(III)
should also be included. See README for more information. Defaults to FALSE. \link{logical}}

\item{wavlength}{User input of the wavelength to return data for. \link{numeric}}
}
\value{
This function returns a table of SUVA values at the specified wavelength.
}
\description{
This function calculates the specific ultra-violet absorbance
(SUVA) for a user-specified wavelength from National Ecological Observatory
Network (NEON) water chemistry data.
}
\examples{
#Using an example file
#outputData <- calcSuvaX(
#absorbanceData=swc_externalLabAbsorbanceScan,
#concentrationData=swc_externalLabDataByAnalyte,
#wavelength=401,
#correctFe=FALSE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Robert Hensley \email{hensley@battelleecology.org} \cr
}
\keyword{Absorbance}
\keyword{DOC,}
\keyword{NEON,}
\keyword{SUVA,}
\keyword{UV}
