% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatWide.R
\name{formatWide}
\alias{formatWide}
\title{formatWide}
\usage{
formatWide(absorbanceData)
}
\arguments{
\item{absorbanceData}{User input of the table of NEON absorbance data \link{data.frame}}
}
\value{
This function returns a table absorbance values in wide format.
}
\description{
This function re-formats National Ecological Observatory
Network (NEON) UV-Vis absorbance data from long format into wide format.
}
\examples{
#Using an example file
#outputData <- formatWide(
#absorbanceData=swc_externalLabAbsorbanceScan)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Robert Hensley \email{hensley@battelleecology.org} \cr
}
\keyword{Absorbance}
\keyword{DOC,}
\keyword{NEON,}
\keyword{SUVA,}
\keyword{UV}
